backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
description = "Backend for Project Management MVP"
authors = [{ name = "placeholder", email = "you@example.com" }]

# using uv package manager
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "openai",
    "pytest",
    "pytest-mock",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.uv]
